refactor: menu_group 을 menu 패키지 내부로 이동 - menu_group 은 menu 와 항상 함께 사용된다.
