+++ src/main/java/kitchenpos/menu/domain/menu/Menu.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.domain.menu;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
@@ -10,7 +10,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
-import kitchenpos.menu_group.domain.MenuGroup;
+import kitchenpos.menu.domain.menu_group.MenuGroup;
 
 import java.math.BigDecimal;
 import java.util.List;
