+++ src/main/java/kitchenpos/menu/domain/menu/MenuProduct.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.domain.menu;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
