+++ src/main/java/kitchenpos/order/delivery_order/application/DeliveryOrderService.java
@@ -4,8 +4,8 @@ import kitchenpos.order.delivery_order.domain.DeliveryOrderRepository;
 import kitchenpos.order.delivery_order.domain.DeliveryOrder;
 import kitchenpos.order.delivery_order.domain.KitchenridersClient;
 import kitchenpos.order.delivery_order.domain.OrderStatus;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.domain.menu.Menu;
+import kitchenpos.menu.domain.menu.MenuRepository;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderType;
 import org.springframework.stereotype.Service;
