+++ src/test/java/kitchenpos/Fixtures.java
@@ -3,9 +3,9 @@ package kitchenpos;
 import kitchenpos.order.delivery_order.domain.OrderStatus;
 import kitchenpos.order.delivery_order.domain.DeliveryOrder;
 import kitchenpos.order.eat_in_order.domain.EatInOrder;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu_group.domain.MenuGroup;
+import kitchenpos.menu.domain.menu.Menu;
+import kitchenpos.menu.domain.menu.MenuProduct;
+import kitchenpos.menu.domain.menu_group.MenuGroup;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderType;
 import kitchenpos.order_table.domain.OrderTable;
