+++ src/test/java/kitchenpos/product/service/ProductServiceTest.java
@@ -3,8 +3,8 @@ package kitchenpos.product.service;
 import kitchenpos.menu.infra.InMemoryMenuRepository;
 import kitchenpos.product.infra.InMemoryProductRepository;
 import kitchenpos.common.domain.FakePurgomalumClient;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.domain.menu.Menu;
+import kitchenpos.menu.domain.menu.MenuRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.common.domain.PurgomalumClient;
